float to int for number weeks

+++ classroom/models.py
@@ -17,7 +17,7 @@ class Semester(models.Model):
     def get_number_weeks(self):
         monday1 = (self.date_begin - timedelta(days=self.date_end.weekday()))
         monday2 = (self.date_end - timedelta(days=self.date_end.weekday()))        
-        return (monday2 - monday1).days / 7
+        return int((monday2 - monday1).days / 7)
 
     def get_year(self):
         pass
